Some cleanup of the BayesianFactor interface.
