+++ src/main/java/ch/idsia/crema/factor/bayesian/BayesianFactor.java
@@ -34,6 +34,7 @@ public interface BayesianFactor extends OperableFactor<BayesianFactor>, Separate
 
 	double getLogValueAt(int index);
 
+	// TODO: this method should only for internal use
 	double[] getData();
 
 	@Override
@@ -54,21 +55,19 @@ public interface BayesianFactor extends OperableFactor<BayesianFactor>, Separate
 	// TODO: these methods below need more consideration on what to do with them and where to put them
 	ObservationBuilder sample();
 
-	double KLDivergence(BayesianFactor approx);
-
 	BayesianFactor replace(double value, double replacement);
 
 	BayesianFactor replaceNaN(double replacement);
 
 	BayesianFactor scale(double k);
 
+	BayesianFactor addition(BayesianFactor factor);
+
 	BayesianFactor reorderDomain(Strides newStrides);
 
 	BayesianFactor reorderDomain(int... vars);
 
 	void sortDomain();
-
-	BayesianFactor addition(BayesianFactor factor);
 
 	double logProb(TIntIntMap[] data, int leftVar);
